Finally solved day 18B!
[–]DarksteelPenguin 8 points 10 hours ago
Spent more than an hour trying to understand what was going wrong with
my solution, which seemed to run an infinite loop. Tried to decompose my
input to understand what is going on. Found an infinite loop. Checked on
reddit to see what the deal is. Found someone explaining that the input
is performing a bubble sort. Realized that jgz requires a value greater
than zero and not different from zero. Changed a character in my code.
Got the correct answer in one go. Cried.